% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{max_abs_xyz}
\alias{max_abs_xyz}
\alias{dist_euclidean}
\alias{rotate}
\title{Utilities for coordinate values}
\usage{
max_abs_xyz(.tbl)

dist_euclidean(.tbl, point = c(x = 0, y = 0, z = 0))

rotate(.tbl, theta, axis = c("z", "x", "y"))
}
\arguments{
\item{.tbl}{data.frame with (x, y, z) columns}

\item{point}{named vector or tibble}

\item{theta}{radian angle}

\item{axis}{a string}
}
\value{
max(abs(x, y, z))

numeric vector

modified data.frame
}
\description{
\code{max_abs_xyz} extracts max values for plot limits.

\code{dist_euclidean} calculates distance from a specified cell.

\code{rotate} modifies coordinates centering on a specified axis.
}
\examples{
max_abs_xyz(data.frame(x=2, y=-3, z=4))
}
