% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{as_cuboid}
\alias{as_cuboid}
\alias{as.data.frame.cuboid}
\alias{as.array.cuboid}
\title{Binary expression of cell existence}
\usage{
as_cuboid(x, expand = 0L)

\method{as.data.frame}{cuboid}(x, row.names = NULL, optional = FALSE, ...)

\method{as.array}{cuboid}(x, ...)
}
\arguments{
\item{x}{a data.frame with (x, y, z) or 3D array.}

\item{expand}{an integer.}

\item{row.names, optional, ...}{ignored; to suppress warnings for S3 methods.}
}
\description{
\code{as_cuboid()} expand a coordinate set to a binary cuboid with zero-padding.
The data.frame representation is ordered in the same way as the array representation.
They have some extra attributes to ensure round-trip conversion.
}
