% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{extract_demography}
\alias{extract_demography}
\title{Extract demography from raw population data}
\usage{
extract_demography(population)
}
\arguments{
\item{population}{data.frame}
}
\description{
Extract demography from raw population data
}
