% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{dist_genealogy}
\alias{dist_genealogy}
\alias{dist_vaf}
\alias{pairwise_distances}
\title{Calculate genetic distance from genealogy or VAF}
\usage{
dist_genealogy(graph, cell_ids)

dist_vaf(vaf, ncell = Inf)

pairwise_distances(graph, regions)
}
\arguments{
\item{graph}{igraph}

\item{cell_ids}{id column of sampled regions}

\item{vaf}{data.frame or matrix. e.g., output from \code{\link[=tally_vaf]{tally_vaf()}}.}

\item{ncell}{number of cells per specimen}

\item{regions}{output of \code{sample_uniform_regions()}}
}
\description{
Calculate genetic distance from genealogy or VAF
}
\details{
\code{dist_genealogy} summarizes branch lengths within/between regions.

\code{dist_vaf} summarizes diversity within/between regions.

\code{pairwise_distances} summarizes physical and genetic distances.
}
\seealso{
\code{\link[=fst]{fst()}}
}
