% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{propagate_drivers}
\alias{propagate_drivers}
\title{Functions to deal with driver mutations}
\usage{
propagate_drivers(drivers, graph)
}
\arguments{
\item{drivers}{data.frame}

\item{graph}{igraph}
}
\description{
\code{propagate_drivers} calculates the effect of drivers in descendant cells
}
