% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{math_score}
\alias{math_score}
\alias{detectable_mutants_all}
\alias{detectable_mutants}
\title{Calculate MATH (mutant-allele tumor heterogeneity) score from variant allele frequencies}
\usage{
math_score(x, constant = 1.4826, na.rm = FALSE)

detectable_mutants_all(population, threshold)

detectable_mutants(graph, nodes, threshold)
}
\arguments{
\item{x}{a numeric vector.}

\item{constant}{scale factor.}

\item{na.rm}{if \code{TRUE} then \code{NA} values are stripped
    from \code{x} before computation takes place.}

\item{population}{tibble with descendants and id column}

\item{threshold}{lowerbound of detectable allele frequency}

\item{graph}{igraph}

\item{nodes}{igraph vertices}
}
\value{
numeric

integer IDs
}
\description{
Calculate MATH (mutant-allele tumor heterogeneity) score from variant allele frequencies

Extract cells whose expected allele frequencies are above threshold
}
