% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{math_score}
\alias{math_score}
\title{Evaluate heterogeneity}
\usage{
math_score(x, constant = 1.4826, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{constant}{scale factor.}

\item{na.rm}{if \code{TRUE} then \code{NA} values are stripped
    from \code{x} before computation takes place.}
}
\value{
numeric
}
\description{
\code{math_score} calculates MATH (mutant-allele tumor heterogeneity) score
from variant allele frequencies.
}
