% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{trans_coord_hex}
\alias{trans_coord_hex}
\title{Transform coordinates into hexagonal grid}
\usage{
trans_coord_hex(.tbl)
}
\arguments{
\item{.tbl}{data.frame with (x, y, z) columns}
}
\description{
Transform coordinates into hexagonal grid
}
