% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_genealogy}
\alias{layout_genealogy}
\title{Layout cell genealogy}
\usage{
layout_genealogy(population)
}
\arguments{
\item{population}{tbl}
}
\description{
Layout cell genealogy
}
\details{
\code{layout_genealogy} returns coordinates of nodes and edges for plotting.
}
