% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{detect_surface}
\alias{detect_surface}
\alias{get_se}
\title{Extract surface cells with mathematical morphology}
\usage{
detect_surface(.tbl, se)

get_se(coord = c("moore", "neumann", "hex"), dimensions = 3L)
}
\arguments{
\item{.tbl}{a data.frame with (x, y, z) columns}

\item{se}{structuring element as a binary array}

\item{coord}{a string}

\item{dimensions}{an integer}
}
\value{
tibble with surface column
}
\description{
\code{detect_surface} adds a binary column

\code{get_se} construct a structuring element (kernel) as a binary array
}
