% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_capture_rate}
\alias{plot_capture_rate}
\title{Plot capture_rate ~ nsam of biopsy}
\usage{
plot_capture_rate(data, point = TRUE, errorbar = TRUE)
}
\arguments{
\item{data}{tbl from \code{summarize_capture_rate} or \code{evaluate_mrs}}

\item{point}{logical}

\item{errorbar}{logical}
}
\description{
Plot capture_rate ~ nsam of biopsy
}
