% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-genealogy.R
\name{augment_genealogy}
\alias{augment_genealogy}
\alias{gggenealogy}
\title{Visualize cell genealogy}
\usage{
augment_genealogy(graph, layout = igraphlite::layout_reingold_tilford)

gggenealogy(data, mapping = ggplot2::aes(), ...)
}
\arguments{
\item{graph}{igraph}

\item{layout}{A function or resulting data.frame.
If not provided, \link[igraphlite]{layout_nicely} is applied.}

\item{data}{tbl from \code{augment_genealogy()}}

\item{mapping}{\code{ggplot2::aes()}}

\item{...}{passed to \code{ggplot2::geom_segment()}}
}
\description{
Visualize cell genealogy
}
\details{
\code{augment_genealogy} calculates coordinates of nodes and edges for plotting.

\code{gggenealogy()} creates a basic ggplot object.
}
