% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{histogram_freqspec}
\alias{histogram_freqspec}
\alias{plot_lattice2d}
\title{ggplot for frequency spectrum}
\usage{
histogram_freqspec(freqs)

plot_lattice2d(
  .tbl,
  colour = "z",
  alpha = 1,
  size = 1,
  limit = max_abs_xyz(.tbl)
)
}
\arguments{
\item{freqs}{numeric vector}

\item{.tbl}{tbl with extant cells}

\item{colour}{column name to colorcode}

\item{alpha}{opacity \verb{[0, 1]}}

\item{size}{relative size of points}

\item{limit}{for value range}
}
\description{
ggplot for frequency spectrum

ggplot for 2D lattice
}
