% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{modify_population}
\alias{modify_population}
\alias{filter_extant}
\alias{filter_common_ancestors}
\title{Functions to modify population data.frame}
\usage{
modify_population(population, coord, dimensions, ..., num_clades = 4L)

filter_extant(population)

filter_common_ancestors(population, threshold = 0.05)
}
\arguments{
\item{population}{tibble}

\item{coord}{string}

\item{dimensions}{integer}

\item{...}{ignored}

\item{num_clades}{integer}

\item{threshold}{minimum frequency of detectable alleles}
}
\description{
\code{modify_population} add various columns to the raw population data

\code{filter_extant} collects the extant cells at the end of the simulation

\code{filter_common_ancestors} collects major common ancestors
}
