% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_confs}
\alias{read_confs}
\alias{read_populations}
\alias{read_results}
\alias{read_snapshots}
\title{Read TSV files in given directories}
\usage{
read_confs(indirs = getwd())

read_populations(indirs = getwd())

read_results(indirs = getwd())

read_snapshots(indirs = getwd())
}
\arguments{
\item{indirs}{a string vector}
}
\description{
\code{read_confs} reads config files

\code{read_populations} reads populations

\code{read_results} reads confs and populations as a nested tibble

\code{read_snapshots} calls \code{read_results} and reads snapshots
}
