% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-combn.R
\name{combn_ids}
\alias{combn_ids}
\alias{summarize_capture_rate}
\title{Sample cells from a population}
\usage{
combn_ids(x, m = seq_along(x))

summarize_capture_rate(combinations, population, threshold = 0.01)
}
\arguments{
\item{x}{list of ID vectors}

\item{m}{number of regions to sample}

\item{combinations}{nested tibble from \code{combn_ids}}

\item{population}{tibble}

\item{threshold}{minimum frequency of detectable alleles}
}
\value{
tibble
}
\description{
\code{combn_ids} makes combinations of cell IDs for various number of samples.

\code{summarize_capture_rate} calculates expected allele capture rate on various combinations of samples.
}
