% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_bulk}
\alias{sample_bulk}
\alias{sample_uniform_regions}
\alias{sample_random_regions}
\title{Sample cells from a population}
\usage{
sample_bulk(tbl, center = c(x = 0, y = 0, z = 0), ncell = 10L)

sample_uniform_regions(tbl, nsam = 2L, ncell = 10L)

sample_random_regions(tbl, nsam = 2L, ncell = 10L)
}
\arguments{
\item{tbl}{tibble of extant cells with id, x, y, z columns}

\item{center}{named (x, y, z) vector, list, or tibble}

\item{ncell}{number of cells per specimen}

\item{nsam}{number of regions to sample}
}
\description{
\code{sample_bulk} samples bulk of cells near the specified center

\code{sample_uniform_regions} sample uniformly distributed regions

\code{sample_random_regions} samples multiple regions at random
}
