% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumopp-package.R
\docType{package}
\name{tumopp-package}
\alias{tumopp-package}
\title{R interface to tumopp, tumor growth simulator in C++}
\description{
R interface to tumopp, tumor growth simulator in C++
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heavywatal/tumopp}
  \item Report bugs at \url{https://github.com/heavywatal/tumopp/issues}
}

}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavy.watal@gmail.com}

}
