% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaf.R
\name{tally_vaf}
\alias{tally_vaf}
\alias{tidy_vaf}
\title{Utilities for variant allele frequency}
\usage{
tally_vaf(samples, sites)

tidy_vaf(tbl)
}
\arguments{
\item{samples}{list of integer IDs}

\item{sites}{list of integer IDs}

\item{tbl}{output of \code{tally_vaf}}
}
\description{
\code{tally_vaf} evaluates overlap of sampled and mutated cells

\code{tidy_vaf} transforms vaf table
}
