% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_results}
\alias{write_results}
\title{Utilities for writing files}
\usage{
write_results(results)
}
\arguments{
\item{results}{data.frame returned from tumopp()}
}
\description{
Utilities for writing files
}
\details{
\code{write_results} writes nested data.frame to files
}
