% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch-length.R
\name{mean_branch_length}
\alias{mean_branch_length}
\alias{within_between_samples}
\title{Calculate genetic distance with igraph}
\usage{
mean_branch_length(graph, from = graph$sink, to = from)

within_between_samples(graph, regions)
}
\arguments{
\item{graph}{igraph}

\item{from, to}{igraph vertices (not cell ID)}

\item{regions}{output of \code{sample_uniform_regions()}}
}
\description{
Calculate genetic distance with igraph
}
\details{
\code{mean_branch_length} calculates mean branch length within/between sub-graphs.

\code{within_between_samples} summarizes branch lengths.
}
