% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{extract_history}
\alias{extract_history}
\alias{summarize_demography}
\title{Functions for event history}
\usage{
extract_history(population)

summarize_demography(history)
}
\arguments{
\item{population}{data.frame}

\item{history}{data.frame from \code{\link[=extract_history]{extract_history()}}}
}
\description{
\code{extract_history} extracts birth and death events from raw population data.
}
\details{
\code{summarize_demography} simplifies history.
}
