% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{stats_dispersion}
\alias{stats_dispersion}
\alias{rpois_dispersion}
\alias{tibble_dchisq}
\title{Utilities for dispersion
\code{stats_dispersion()} calculates various dispersion stats.}
\usage{
stats_dispersion(x)

rpois_dispersion(n, lambda, nrep = 1L)

tibble_dchisq(x, n)
}
\arguments{
\item{x}{A numeric vector}

\item{n, lambda}{passed to \code{\link[stats:Poisson]{stats::rpois()}}.}

\item{nrep}{number of replications.}
}
\description{
Utilities for dispersion
\code{stats_dispersion()} calculates various dispersion stats.
}
\details{
\code{rpois_dispersion()} summarizes the results of \code{\link[stats:Poisson]{stats::rpois()}}.

\code{tibble_dchisq()} generates a tibble.
}
