% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{fst}
\alias{fst}
\alias{gst}
\alias{pairwise_fst}
\title{Calculate F_ST from matrix}
\usage{
fst(m, sample_sizes = NULL)

gst(m, sample_sizes = NULL)

pairwise_fst(m, sample_sizes = NULL)
}
\arguments{
\item{m}{matrix output from \code{\link[=dist_genealogy]{dist_genealogy()}} or \code{\link[=dist_vaf]{dist_vaf()}}}

\item{sample_sizes}{cell numbers of each region}
}
\description{
Calculate F_ST from matrix
}
\details{
\code{fst} calculates Hudson's F_ST estimator.
\itemize{
\item theta: Weir and Cockerham (1984); Weir and Hill (2002).
\item N_ST: Lynch and Crease (1990).
\item <F_ST>: Hudson, Slatkin, and Maddison (1992) Genetics.
}

\code{gst} calculates Nei's F_ST estimator.
\itemize{
\item G_ST, gamma_ST: Nei (1973, 1977, 1982).
\item K_ST: Hudson, Boos, and Kaplan (1992) MBE.
}
}
\seealso{
\code{\link[=dist_genealogy]{dist_genealogy()}}
}
