% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{make_igraph}
\alias{make_igraph}
\alias{subtree}
\alias{trim_root}
\alias{internal_nodes}
\title{Functions depending on igraph}
\usage{
make_igraph(population)

subtree(graph, nodes = integer(0L), trim = FALSE)

trim_root(graph)

internal_nodes(graph, nodes, sensitivity)
}
\arguments{
\item{population}{tbl}

\item{graph}{igraph}

\item{nodes}{integer cell IDs}

\item{trim}{whether to remove common ancestors older than MRCA.}

\item{sensitivity}{minimum allele frequency}
}
\description{
Functions depending on igraph
}
\details{
\code{make_igraph} converts raw population tbl into graph.

\code{subtree} extracts subgraph among terminal nodes.

\code{trim_root()} removes ancestral nodes older than MRCA.

\code{internal_nodes} selects major common ancestors above threshold.
}
