% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{make_igraph}
\alias{make_igraph}
\alias{subtree}
\alias{internal_nodes}
\title{Functions depending on igraph}
\usage{
make_igraph(population)

subtree(graph, nodes = integer(0L))

internal_nodes(graph, nodes, sensitivity)
}
\arguments{
\item{population}{tbl}

\item{graph}{igraph}

\item{nodes}{integer cell IDs}

\item{sensitivity}{minimum allele frequency}
}
\description{
\code{make_igraph} converts raw population tbl into graph.

\code{subtree} extracts subgraph among terminal nodes.

\code{internal_nodes} selects major common ancestors above threshold.
}
