% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{add_surface}
\alias{add_surface}
\title{Extract surface cells with mathematical morphology}
\usage{
add_surface(population, coord, dimensions)
}
\arguments{
\item{population}{a data.frame with (x, y, z) before hex transformation}

\item{coord}{a string}

\item{dimensions}{an integer}
}
\description{
\code{add_surface} adds a binary column to the data.frame
}
