% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.R
\name{mutate_clades}
\alias{mutate_clades}
\title{Sprinkle mutations on genealogy}
\usage{
mutate_clades(graph, mu = NULL, segsites = NULL)
}
\arguments{
\item{graph}{igraph}

\item{mu}{mutation rate per cell division (ignored if segsites is given)}

\item{segsites}{number of segregating sites}
}
\description{
\code{mutate_clades} samples nodes to mutate and returns their descendants.
}
