% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.R
\name{make_sample}
\alias{make_sample}
\title{Sprinkle mutations on genealogy}
\usage{
make_sample(graph, nsam = 0L, mu = NULL, segsites = NULL)
}
\arguments{
\item{graph}{igraph}

\item{nsam}{number of cells to sample}

\item{mu}{mutation rate per cell division (ignored if segsites is given)}

\item{segsites}{number of segregating sites}
}
\description{
Sprinkle mutations on genealogy
}
\details{
\code{make_sample} creates a genotype matrix from a given genealogy tree.
}
