% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_capture_rate}
\alias{plot_capture_rate}
\title{Plot capture_rate ~ nsam of biopsy}
\usage{
plot_capture_rate(data, point = 1, alpha = 0.3, errorbar = TRUE)
}
\arguments{
\item{data}{tbl from \code{\link[=summarize_capture_rate]{summarize_capture_rate()}} or \code{\link[=evaluate_mrs]{evaluate_mrs()}}}

\item{point}{size of a data point}

\item{alpha}{opacity of a data point}

\item{errorbar}{logical}
}
\description{
Plot capture_rate ~ nsam of biopsy
}
