% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rgl.R
\name{plot_tumor3d}
\alias{plot_tumor3d}
\alias{snapshot_surface}
\alias{add_col}
\title{3D plotting with rgl}
\usage{
plot_tumor3d(.tbl = NULL, limits = NULL)

snapshot_surface(.tbl, filename = tempfile("rgl_", fileext = ".png"), ...)

add_col(.tbl, column = "clade", palette = "Spectral", direction = 1)
}
\arguments{
\item{.tbl}{data.frame with (x, y, z, col)}

\item{limits}{passed to xlim, ylim, zlim of \code{rgl::plot3d()}}

\item{filename}{string}

\item{...}{passed to \code{plot_tumor3d()}}

\item{column}{column name to colorcode}

\item{palette}{name of ColorBrewer palette}

\item{direction}{-1 to reverse color scale}
}
\description{
3D plotting with rgl
}
\details{
\code{plot_tumor3d} plots tumor in 3D with rgl.

\code{snapshot_surface} is a shortcut of \code{plot_tumor3d()} and \code{rgl::snapshot3d()}.

\code{add_col} adds a column for color in \code{plot_tumor3d()}.
}
