% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rgl.R
\name{plot_tumor3d}
\alias{plot_tumor3d}
\alias{snapshot_surface}
\title{3D plotting with rgl}
\usage{
plot_tumor3d(.tbl, colour = "clade", .palette = "Spectral",
  .reverse = FALSE, .min = NULL)

snapshot_surface(.tbl, filename = tempfile("rgl_", fileext = ".png"),
  ...)
}
\arguments{
\item{.tbl}{data.frame}

\item{colour}{column name to colorcode}

\item{.palette}{name for RColorBrewer::brewer.pal()}

\item{.reverse}{logical for order of color vector}

\item{.min}{minimum limit of axes}

\item{filename}{string}

\item{...}{passed to plot_tumor3d()}
}
\description{
\code{plot_tumor3d} plots tumor in 3D with rgl

\code{snapshot_surface} is a shortcut of \code{plot_tumor3d} and \code{snapshot3d}
}
