% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{modify_population}
\alias{modify_population}
\alias{filter_extant}
\title{Modify population table}
\usage{
modify_population(population, coord, dimensions, ..., num_clades = 4L)

filter_extant(population)
}
\arguments{
\item{population}{tibble}

\item{coord}{string}

\item{dimensions}{integer}

\item{...}{ignored}

\item{num_clades}{integer}
}
\description{
Modify population table

Filter extant cells
}
