% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{filter_extant}
\alias{filter_extant}
\alias{count_extant}
\alias{filter_common_ancestors}
\title{Functions to modify population data.frame}
\usage{
filter_extant(population)

count_extant(population)

filter_common_ancestors(population, threshold = 0.05)
}
\arguments{
\item{population}{tibble}

\item{threshold}{minimum frequency of detectable alleles}
}
\description{
\code{filter_extant} collects the extant cells at the end of the simulation.

\code{count_extant} counts the number of extant cells

\code{filter_common_ancestors} collects major common ancestors.
}
