% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_confs}
\alias{read_confs}
\alias{read_results}
\alias{read_tumopp}
\title{Read TSV files in given directories}
\usage{
read_confs(indirs = getwd(), mc.cores = getOption("mc.cores", 1L))

read_results(
  indirs = getwd(),
  mc.cores = getOption("mc.cores", 1L),
  graph = TRUE
)

read_tumopp(file)
}
\arguments{
\item{indirs}{a string vector}

\item{mc.cores}{The number of cores to use for concurrent execution.}

\item{graph}{add graph column if TRUE}

\item{file}{passed as the first argument of \code{\link[readr:read_delim]{readr::read_tsv()}}}
}
\description{
\code{read_confs} reads config files.
}
\details{
\code{read_results} reads configs and populations as a nested tibble.

\code{read_tumopp} is an alias of \code{\link[readr:read_delim]{readr::read_tsv()}} with some options
to read result files: \code{population.tsv.gz} and \code{snapshots.tsv.gz}.
}
