% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_uniform_regions}
\alias{sample_uniform_regions}
\alias{sample_random_regions}
\alias{sample_bulk}
\alias{evaluate_mrs}
\title{Sample cells from a population}
\usage{
sample_uniform_regions(tbl, nsam = 2L, ncell = 10L, jitter = 0)

sample_random_regions(tbl, nsam = 2L, ncell = 10L)

sample_bulk(tbl, center = c(x = 0, y = 0, z = 0), ncell = 10L)

evaluate_mrs(population, nsam, ncell, threshold = 0.05)
}
\arguments{
\item{tbl}{tibble of extant cells with id, x, y, z columns}

\item{nsam}{number of regions to sample}

\item{ncell}{number of cells per specimen}

\item{jitter}{amount of random variations on x and y axes}

\item{center}{named (x, y, z) vector, list, or tibble}

\item{population}{tbl}

\item{threshold}{minimum frequency of detectable alleles}
}
\description{
\code{sample_uniform_regions} sample uniformly distributed regions

\code{sample_random_regions} samples multiple regions at random

\code{sample_bulk} samples a bulk of cells near the specified center

\code{evaluate_mrs} is a shortcut to evaluate multi-region sampling
}
