% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{genetic_stats}
\alias{genetic_stats}
\alias{morphological_stats}
\title{Summarizing functions}
\usage{
genetic_stats(extant)

morphological_stats(extant)
}
\arguments{
\item{extant}{tibble}
}
\value{
tibble

tibble
}
\description{
\code{genetic_stats} calculates summary statistics of genealogy

\code{morphological_stats} calculates summary statistics of morphology
}
