% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumopp-package.R
\docType{package}
\name{tumopp-package}
\alias{tumopp-package}
\title{tumopp: R interface to tumopp, a tumor growth simulator in C++}
\description{
R interface to tumopp, a tumor growth simulator in C++.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://heavywatal.github.io/rtumopp}
  \item \url{https://github.com/heavywatal/rtumopp}
  \item Report bugs at \url{https://github.com/heavywatal/rtumopp/issues}
}

}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavy.watal@gmail.com} (\href{https://orcid.org/0000-0002-4718-7605}{ORCID})

}
\keyword{internal}
