% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumopp.R
\name{tumopp}
\alias{tumopp}
\alias{mslike}
\alias{make_args}
\title{Run tumopp}
\usage{
tumopp(args = character(0L), nsam = 0L)

mslike(nsam = 20L, args = character(0L))

make_args(alt, const = NULL, nreps = 1L)
}
\arguments{
\item{args}{command line arguments as a string vector or list of strings}

\item{nsam}{number of samples for ms-like output}

\item{alt}{named list of altered arguments}

\item{const}{unnamed vector of constant arguments}

\item{nreps}{number of repeats}
}
\description{
\code{tumopp()} returns full results with config columns in a data.frame.

\code{mslike()} returns only binary genotypes in ms-like format.

\code{make_args()} returns argument combinations in a list.
}
