% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumopp.R
\name{tumopp}
\alias{tumopp}
\alias{tumopp.default}
\alias{tumopp.list}
\alias{tumopp.data.frame}
\alias{make_args}
\title{Run tumopp}
\usage{
tumopp(args, ...)

\method{tumopp}{default}(args = character(0L), ..., graph = TRUE, cache = FALSE)

\method{tumopp}{list}(args, ..., graph = TRUE, mc.cores = getOption("mc.cores", 1L))

\method{tumopp}{data.frame}(args, ..., graph = TRUE, mc.cores = getOption("mc.cores", 1L))

make_args(alt, const = NULL, times = 1L, each = 1L)
}
\arguments{
\item{args}{command line arguments as a string vector or list of strings.}

\item{...}{not used.}

\item{graph}{add graph column if TRUE}

\item{cache}{use cache if TRUE.}

\item{mc.cores}{The number of cores to use for concurrent execution.}

\item{alt}{named list of altered arguments.}

\item{const}{named list of constant arguments.}

\item{times, each}{passed to \code{\link[=rep]{rep()}}}
}
\description{
\code{tumopp()} returns full results with config columns in a data.frame.
See \code{tumopp("-h")} or \url{https://heavywatal.github.io/tumopp/group__params.html}
for the list of command options.
}
\details{
A \code{population} data.frame includes ancestral cells.
Extant cells can be extracted by filtering with \code{death == 0} or \code{\link[=filter_extant]{filter_extant()}} function.
The sampling time, i.e., the end of a simulation, is typically the maximum value of birth time.

The default unit of time (\code{birth} and \code{death} columns) is the average cell cycle of newborn cells
(given the parameter \code{-b}/\code{--beta0} is set to 1).
For example, step-wise tumor growth and integer values in the birth column will be observed
if \code{-k}/\code{--shape} parameter is set to a very large value like \code{10**6}.
If you are considering some cell line whose average cell cycle is 4 days for example,
then the unit of those columns can be interpreted as 4 days,
or you can set \code{--beta0=0.25} to change the unit to a day.

The \code{omega} column denotes the number of cell devisions allowed for each cell.
Negative values denote unlimited proliferation potential.

\code{make_args()} returns argument combinations in a tibble.
}
