% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaf.R
\name{make_vaf}
\alias{make_vaf}
\alias{tally_vaf}
\alias{tidy_vaf}
\alias{filter_detectable}
\alias{sort_vaf}
\alias{dist_vaf}
\title{Utilities for variant allele frequency}
\usage{
make_vaf(graph, samples, mu, threshold = 0.05)

tally_vaf(samples, sites)

tidy_vaf(tbl)

filter_detectable(tbl, threshold)

sort_vaf(tbl, method = c("average", "ward.D2", "complete", "single"))

dist_vaf(tbl, ncell = Inf)
}
\arguments{
\item{graph}{igraph}

\item{samples}{list of integer IDs}

\item{mu}{mutation rate per cell division (ignored if segsites is given)}

\item{threshold}{minimum detectable frequency}

\item{sites}{list of integer IDs}

\item{tbl}{output of \code{tally_vaf}}

\item{method}{passed to \code{stats::hclust}}

\item{ncell}{number of cells per specimen}
}
\description{
Utilities for variant allele frequency
}
\details{
\code{make_vaf} is a shortcut to make neutral VAF pattern.

\code{tally_vaf} evaluates overlap of sampled and mutated cells.

\code{tidy_vaf} transforms vaf table.

\code{filter_detectable} removes sites where freq < threshold.

\code{sort_vaf} reorders rows and columns of VAF table.

\code{dist_vaf} summarizes diversity within/between regions.
}
\seealso{
\code{\link[=dist_genealogy]{dist_genealogy()}}
}
