% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaf.R
\name{make_vaf}
\alias{make_vaf}
\alias{make_longer_vaf}
\alias{longer_vaf}
\alias{filter_detectable}
\alias{sort_vaf}
\title{Utilities for variant allele frequency}
\usage{
make_vaf(graph, samples, mu)

make_longer_vaf(graph, samples, mu, threshold = 0.05)

longer_vaf(vaf)

filter_detectable(vaf, threshold)

sort_vaf(vaf, method = c("average", "ward.D2", "complete", "single"))
}
\arguments{
\item{graph}{igraph}

\item{samples}{list of integer IDs}

\item{mu}{mutation rate per cell division (ignored if segsites is given)}

\item{threshold}{minimum detectable frequency}

\item{vaf}{output of \code{\link[=make_vaf]{make_vaf()}}}

\item{method}{passed to \code{stats::hclust}}
}
\description{
Utilities for variant allele frequency
}
\details{
\code{make_vaf} generates neutral VAF pattern from cell genealogy.

\code{make_longer_vaf} is a shortcut to make VAF in longer format.

\code{longer_vaf} transforms vaf table.

\code{filter_detectable} removes sites where freq < threshold.

\code{sort_vaf} reorders rows and columns of VAF table.
}
